+++ my_music_code/lib/Auth/Service/auth_service.dart
@@ -18,15 +18,6 @@ Map<String, String> errorMap = {
 };
 
 
-
-class SignLoginModel {
-  String username = "";
-  String email = "";
-  String password = "";
-  String confirmPassword = "";
-  bool creatingAccount = false;
-}
-
 class AuthService {
   Future<void> controlSignLogin({
     required BuildContext context,
