+++ my_music_code/lib/Models/sign_login_model.dart
@@ -1,0 +1,7 @@
+class SignLoginModel {
+  String username = "";
+  String email = "";
+  String password = "";
+  String confirmPassword = "";
+  bool creatingAccount = false;
+}
