+++ my_music_code/lib/MyPlaylists/playlist_page.dart
@@ -4,14 +4,6 @@ import 'package:my_music_code/Globals/responsive_text.dart';
 import 'package:my_music_code/Globals/size_config.dart';
 import 'package:my_music_code/Globals/style.dart';
 
-class PlaylistModel{
-  final String name;
-  final String image;
-  final String creator;
-  final int songs;
-  PlaylistModel({this.name = "Playlist Name",this.image = DefaultPlaceholder.image, this.creator = "creator", this.songs = 40});
-}
-
 class MyPlaylistPage extends StatefulWidget {
   const MyPlaylistPage({super.key});
   @override
