+++ my_music_code/lib/universal.dart
@@ -2,8 +2,8 @@ library universal;
 
 import 'package:audioplayers/audioplayers.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:my_music_code/Auth/Service/auth_service.dart';
 import 'package:my_music_code/Models/music_model.dart';
+import 'package:my_music_code/Models/sign_login_model.dart';
 import 'package:spotify/spotify.dart' hide User;
 
 AudioPlayer audioPlayer = AudioPlayer();
